requirements.txt
Flask==3.0.3
mysql_connector_repackaged==0.3.1
Pillow==10.3.0
pyzbar==0.1.9
qrcode==7.4.2
Werkzeug==3.0.2
